BPE26-ref: extract the original player operations to dedicated file
