+++ src/features/ui/index.ts
@@ -1,1 +1,2 @@
 export { createFullscreenButtonSection } from "./fullscreen-button";
+export { setupPlayerStickiness } from "./player-stickiness";
